Added animation, game over screen, damage towards bodys It was added animation for when the ship is moving, although it isnt  very good, or very telling of the direction the player is taking
It was added a game over screen, to make for easy testing and replay-ability.

Implemented new interface that allows bodys who implement it to take damage from the enviroment.
Also it lets them be "killed" and will allow for interactions in between the ship and planets/stars.
And between the planets and the stars.

I will have to add  damage for collision and fix some of the animation ui, to have the project as a better prototype.